requirements.txt
numpy
pandas
opencv-python
pysimplegui
PyYAML
tables
scikit-image
easydict